Avoid crash in thread.inc.php when running huge expire.

+++ Rocksolid_Light/rslight/scripts/expire.php
@@ -76,7 +76,9 @@ file_put_contents($logfile, "\n".format_log_date()." ".$config_name." ".$group."
         echo "Expiring: ".$break[4]." IN: ".$group." #".$break[0]."\r\n";
         file_put_contents($logfile, "\n".format_log_date()." ".$config_name." ".$group." Expiring: ".$break[4], FILE_APPEND);
       // Remove article from tradspool:  
-        unlink($spooldir.'/articles/'.$grouppath.'/'.$break[0]);
+        if(is_file($spooldir.'/articles/'.$grouppath.'/'.$break[0])) {
+            unlink($spooldir.'/articles/'.$grouppath.'/'.$break[0]);
+        }
 		thread_cache_removearticle($group,$break[4]);
         continue;
       } else {
